feat: move upload file logic to CsvContext
