+++ src/components/UploadPage.tsx
@@ -1,85 +1,13 @@
 "use client";
 
-import React, { useContext, useState } from "react";
-import Papa from "papaparse";
-import { Box, Modal, Typography, Button } from "@mui/material";
-import { DataGrid, GridColDef } from "@mui/x-data-grid";
+import React, { useContext } from "react";
+import { DataGrid } from "@mui/x-data-grid";
 import { CsvContext } from "../lib/CsvContext";
 import UploadFileIcon from "@mui/icons-material/UploadFile"; // Import an upload icon
 
 const UploadPage: React.FC = () => {
-  const { csvFile, csvData, handleFileLoaded } = useContext(CsvContext);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [file, setFile] = useState<File | null>(null);
-  const [uploadedData, setUploadedData] = useState<Record<string, unknown>[]>(
-    [],
-  );
-  const handleFileUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (file) {
-      // Limit file size under 10mb
-      if (file.size > 10 * 1024 * 1024) {
-        alert("File size limit exceeded. Please choose a smaller file.");
-        return;
-      }
-
-      Papa.parse(file, {
-        header: true,
-        skipEmptyLines: true,
-        complete: (results: { data: Record<string, unknown>[] }) => {
-          // Normalize the CSV data before setting it
-          const normalizedData = normalizeCsvData(results.data);
-
-          // Set the file and normalized data
-          setFile(file);
-          setUploadedData(normalizedData);
-          setIsModalOpen(true); // Open preview modal
-        },
-      });
-    }
-  };
-
-  // Normalize CSV data: convert empty strings to null
-  const normalizeCsvData = (data: Record<string, unknown>[]) => {
-    return data.map((row) => {
-      const normalizedRow: Record<string, unknown> = {};
-
-      Object.keys(row).forEach((key) => {
-        // If the value is an empty string, replace it with null
-        if (row[key] === "") {
-          normalizedRow[key] = null;
-        } else {
-          normalizedRow[key] = row[key];
-        }
-      });
-
-      return normalizedRow;
-    });
-  };
-
-  const handleFileLoad = () => {
-    if (file && uploadedData.length > 0) {
-      handleFileLoaded(file, uploadedData);
-      setIsModalOpen(false); // Close the modal
-    }
-  };
-
-  // Dynamically generate columns for the table preview
-  const columns: GridColDef[] = csvData.length
-    ? Object.keys(csvData[0]).map((key) => ({
-        field: key,
-        headerName: key,
-        flex: 1,
-        minWidth: 150,
-      }))
-    : uploadedData.length
-      ? Object.keys(uploadedData[0]).map((key) => ({
-          field: key,
-          headerName: key,
-          flex: 1,
-          minWidth: 100,
-        }))
-      : [];
+  const { csvFile, csvData, columns, handleFileUpload } =
+    useContext(CsvContext);
 
   return (
     <div className="flex flex-col h-full w-full bg-primary-main text-neutral-white10">
@@ -162,42 +90,6 @@ const UploadPage: React.FC = () => {
           </div>
         </div>
       )}
-
-      <Modal open={isModalOpen} onClose={() => setIsModalOpen(false)}>
-        <Box
-          sx={{
-            position: "absolute",
-            top: "50%",
-            left: "50%",
-            transform: "translate(-50%, -50%)",
-            width: "80%",
-            bgcolor: "background.paper",
-            boxShadow: 24,
-            p: 4,
-          }}
-        >
-          <Typography variant="h6" gutterBottom>
-            CSV File Preview
-          </Typography>
-          <Box sx={{ height: 400, width: "100%" }}>
-            <DataGrid
-              rows={uploadedData.map((row, index) => ({
-                id: index,
-                ...row,
-              }))}
-              columns={columns}
-            />
-          </Box>
-          <Button
-            onClick={handleFileLoad}
-            variant="contained"
-            color="secondary"
-            sx={{ mt: 2 }}
-          >
-            Load
-          </Button>
-        </Box>
-      </Modal>
     </div>
   );
 };
